Fix equinox code

+++ bundles/tools/tool.cli/src/org/eclipse/equinox/internal/cm/reliablefile/ReliableFileInputStream.java
@@ -37,12 +37,12 @@ public class ReliableFileInputStream extends FilterInputStream {
 	/** 
 	 * current position reading from file
 	 */
-	private int readPos;
+	private long readPos;
 
 	/** 
 	 * total file length available for reading
 	 */
-	private int length;
+	private long length;
 
 	/**
 	 * Constructs a new ReliableFileInputStream on the file named <code>name</code>.  If the
@@ -132,7 +132,7 @@ public class ReliableFileInputStream extends FilterInputStream {
 
 		if (num != -1) {
 			if (num + readPos > length) {
-				num = length - readPos;
+				num = (int) (length - readPos);
 			}
 			readPos += num;
 		}
@@ -169,7 +169,7 @@ public class ReliableFileInputStream extends FilterInputStream {
 	 */
 	public synchronized int available() {
 		if (readPos < length) // just in case
-			return (length - readPos);
+			return (int) (length - readPos);
 		return 0;
 	}
